replace hit-transaction-max-spans.js OTel Bridge test case with createSpan-returns-null.js to test the intended OTelTracer.startSpan case where the internal createSpan API returns a null
